factor out test_transfer_matrix into its own test executable
